\name{tabela_klas_r}
\alias{tabela_klas_r}
\title{Klasy kolumn w R}
\usage{
tabela_klas_r(dane)
}
\arguments{
\item{dane}{Ramka danych.}
}
\value{
Tabela nazw kolumn, klas R i typów wewnętrznych.
}
\description{
Klasy kolumn w R
}
\examples{
tabela_klas_r(data.frame(czas = c(2, 4), grupa = c("A", "B")))
}
