\name{oddaj_projekt}
\alias{oddaj_projekt}
\title{Oddanie indywidualnego projektu ilościowego}
\usage{
oddaj_projekt(katalog = ".")
}
\arguments{
\item{katalog}{Katalog projektu.}
}
\value{
Pokwitowanie GitHub z SHA i czasem serwera.
}
\description{
Oddanie indywidualnego projektu ilościowego
}
\examples{
\dontrun{ oddaj_projekt("moje-badania") }
}
