\name{sprawdz_srodowisko}
\alias{sprawdz_srodowisko}
\title{Kontrola środowiska pracy}
\usage{
sprawdz_srodowisko()
}
\value{
Tabela narzędzi, stanów i znaczenia braków.
}
\description{
Nie instaluje pakietów ani nie zmienia konfiguracji. LaTeX jest potrzebny
do budowania PDF, lecz nie do wykonywania zadań lub korzystania z gotowych materiałów.
}
\examples{
sprawdz_srodowisko()
}
